backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pymongo==4.7.0
motor==3.4.0
pydantic==2.11.7
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.6
httpx==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
requests==2.31.0
beautifulsoup4==4.12.3
bcrypt==4.3.0
cryptography==45.0.5
psutil==7.0.0
